# uint256: optimize div-related functions by reducing bounds check