@AaronChen0 2024-05-16
Using godbolt(https://go.godbolt.org/z/4nGefdYrM), I do see the bounds check is moved outside the loop.

If you benchmark using an idle computer (don't benchmark with background jobs running), you definitely would see an upto 7% boost, several `ns` slashed off.

If the percent number in `37.59n ± 1%` is large in benchmark result, that would suggest your computer is busy with other tasks during benchmarking. Try killing them before benchmarking. You could get a more accurate result.

I have another PR doing loop unrolling which combined with this PR could have an upto 17% boost for div-related functions.