@holiman 2024-05-16
> Should I add loop unrolling in this PR to make the performance boost more obvious, or in another PR?

Let's do that in a separate PR.